move sasha files to rewriting folder
